remove sentence from authors notes
Reading through I thought this sentence seemed out of place. It reads as though it is a note from an earlier draft by the author. It doesn't fit the context of the surrounding text which is being able to modify all open source code which is used to evaluate every function in Emacs.

In addition...

- minor punctuation change around incompatible programming languages.

Thanks for writing and sharing this. I'm enjoying it and find it useful.